Refactor move generation and handling for improved clarity and consistency
- Moves are now represented in just 16 bit integer

- Updated move creation to use createMove function across various modules for uniformity.
- Enhanced pawn move generation to exclude the king from opponent pieces during capture checks.
- Modified king check logic to handle cases where the king is not present.
- Refined move comparison logic in search to simplify equality checks.
- Improved UCI command handling for move parsing and validation.
- Adjusted test utilities to align with new move handling functions.

+++ src/core/bitboard.h
@@ -7,7 +7,7 @@
 #include "chess_types.h"
 
 // A1 = LSB, H8 = MSB, little-endian rank-file mapping
-typedef unsigned long long Bitboard;
+typedef uint64_t Bitboard;
 
 /**
  * @brief Checks if a bitboard is empty (i.e., all bits are zero).
